Split MobileMessagingFacade into sender and receiver component classes, rename to MessageReceiverService for clearer parity with original repository
