+++ mobile/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.SimpleTimeTrackerForWearOS"
         tools:targetApi="31">
         <service
-            android:name=".MobileMessagingFacade"
+            android:name=".MessageReceiverService"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
